First pass run at forking askalono as license-store
The goal of license-store is to serve as a support library in a larger
project for automated license audit, review, and approval. The askalono
project has done some great work that we'd like to build on and taylor.

Every effort is being made to respect the Apache 2.0 terms under which
askalono was provided, including preserving the substance of the NOTICE
file, and prominently marking modified files by adding my copyright in
addition to the existing Amazon copyright.

I have also stripped out some bits that I'm not (yet?) interested in
(wasm and cli).

Furthermore, the in-tree spdx dataset was stripped, as I intend to port
this over to using the spdx-dataset crate for that.